Jamf Post Pre-stage Name Mac
Set to trigger after Pre-stage enrollment is complete. This script will grab the currently logged in user to set the Variable 'computerName' in order to prompt the user (and not root) for a new host name for the host mac. Works with 10.14.X(Mojave) and below. Can also be put into self-service without adjustments.